feat: authentication backend

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../model/goalModal");
+const User = require("../model/userModal");
 
 // @desc get a goal
 // @route GET /api/goal
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find();
+	const goals = await Goal.find({ user: req.user.id });
 	res.status(200).json(goals);
 });
 
@@ -18,6 +19,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	});
 	res.status(200).json(goal);
 });
@@ -30,6 +32,18 @@ const updateGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error("goal not found");
+	}
+
+	const user = await User.findById(req.user.id);
+
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+
+	if (user.id !== goal.user.toString()) {
+		res.status(401);
+		throw new Error("User not athorized");
 	}
 
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -47,6 +61,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error("goal not found");
+	}
+
+	const user = await User.findById(req.user.id);
+
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+
+	if (user.id !== goal.user.toString()) {
+		res.status(401);
+		throw new Error("User not athorized");
 	}
 
 	const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
